4/22/20
Did some cleanup of assets.
Tried out added a post processing effect, doesnt seem to be working.
Added a load system, at the moment it only loads the suers username, but in due time it will also load the position and all that jazz.
Moved the pause screen to the Main Canvas so it's not the only thing on that canvas.
Created a few on event scripts to replace some older stuff.